+++ Locales/enUS.lua
@@ -7,6 +7,7 @@ L["Addon Debug Prefix"] = "|cff64b5f6StoryMode Debug:|r "
 L["Addon Legacy Prefix"] = "|cff64b5f6StoryMode:|r "
 
 L["Category Stories"] = "Stories"
+L["Category Active Stories"] = "Active Stories"
 L["Category Epic Stories"] = "Epic Stories"
 L["Category Character Stories"] = "Character Stories"
 L["Category Short Stories"] = "Short Stories"